--- src/main/kotlin/com/wafflytime/api/DemoController.kt
@@ -1,10 +1,0 @@
-package com.wafflytime.api
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class DemoController {
-    @GetMapping("/")
-    fun helloWorld() = "Hello World"
-}
