--- src/main/kotlin/com/wafflytime/database/BaseTimeEntity.kt
@@ -1,25 +1,0 @@
-package com.wafflytime.database
-
-import jakarta.persistence.*
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-import java.time.temporal.ChronoUnit
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener::class)
-abstract class BaseTimeEntity {
-
-    @CreatedDate
-    @Column(columnDefinition = "datetime(6) default '1999-01-01'", updatable = false)
-    var createdAt: LocalDateTime = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS)
-
-    @LastModifiedDate
-    @Column(columnDefinition = "datetime(6) default '1999-01-01'")
-    var modifiedAt: LocalDateTime = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS)
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0L
-}
